Redundant casts

+++ equ-clt/src/main/java/equ/clt/handler/atol/AtolHandler.java
@@ -454,7 +454,7 @@ public class AtolHandler extends DefaultCashRegisterHandler<AtolSalesBatch> {
                             if (dateReceipt == null || startDate == null || dateReceipt.compareTo(startDate) >= 0)
                                 currentSalesInfoList.add(new SalesInfo(false, nppGroupMachinery, numberCashRegister, numberZReport,
                                         dateReceipt, timeReceipt, numberReceipt, dateReceipt, timeReceipt, null, null, null, null,
-                                        null, (BigDecimal) null, barcodeItem, null, itemObject, null, quantityReceiptDetail, priceReceiptDetail,
+                                        null, null, barcodeItem, null, itemObject, null, quantityReceiptDetail, priceReceiptDetail,
                                         sumReceiptDetail, discountSumReceiptDetail, null, null, numberReceiptDetail, file.getName(), null, cashRegister));
                         } else if (isCancelDocument) {
                             cancelReceiptSet.add(numberReceipt);
